alera12/requirements.txt
torch>=1.13.0
torchvision
pandas
numpy
scikit-learn
pillow
tqdm
pydicom
pyyaml
matplotlib
